suggest for those that use versioning after a dash
for our org, we use "verb - app - version" so in its current implementation, grepping the version fails for me. Originally I just changed mine to {print $3), however this caused my version number to be " X.X.X" as opposed to "X.X.X" and that caused issues as well. 

my suggestion in practice so far eliminates the need to specify whether the versioning is locked behind () or -. That said, this is assuming that there's nothing else in the policy name that would match a version number...